# public/openapi.yaml
openapi: 3.0.3
info:
  title: Translation Management API
  description: |
    A scalable API service for managing translations across multiple locales.

    ## Features
    - Store translations for multiple locales (en, fr, es, etc.)
    - Tag translations for context (mobile, desktop, web)
    - Search translations by key, content, or tags
    - Export translations as JSON for frontend applications

    ## Authentication
    All endpoints except `/api/login` require Bearer token authentication.
    Use the login endpoint to obtain a token.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com

servers:
  - url: http://localhost:8080
    description: Docker Local
  - url: http://localhost:8000
    description: Local Development

tags:
  - name: Authentication
    description: User authentication endpoints
  - name: Translations
    description: Translation CRUD and search operations
  - name: Tags
    description: Tag management
  - name: Export
    description: Translation export endpoints

paths:
  /api/login:
    post:
      tags:
        - Authentication
      summary: Login to obtain API token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: admin@example.com
                password:
                  type: string
                  format: password
                  example: password
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  token:
                    type: string
                  user:
                    $ref: "#/components/schemas/User"
        "422":
          description: Validation error

  /api/logout:
    post:
      tags:
        - Authentication
      summary: Logout and revoke token
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Successfully logged out

  /api/translations:
    get:
      tags:
        - Translations
      summary: List all translations
      security:
        - bearerAuth: []
      parameters:
        - name: per_page
          in: query
          schema:
            type: integer
            default: 15
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranslationCollection"
    post:
      tags:
        - Translations
      summary: Create a new translation
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTranslation"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Translation"

  /api/translations/{id}:
    get:
      tags:
        - Translations
      summary: Get a translation by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Translation"
        "404":
          description: Not found
    put:
      tags:
        - Translations
      summary: Update a translation
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTranslation"
      responses:
        "200":
          description: Success
        "404":
          description: Not found
    delete:
      tags:
        - Translations
      summary: Delete a translation
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content
        "404":
          description: Not found

  /api/translations/search:
    get:
      tags:
        - Translations
      summary: Search translations
      security:
        - bearerAuth: []
      parameters:
        - name: key
          in: query
          schema:
            type: string
          description: Filter by translation key pattern
        - name: locale
          in: query
          schema:
            type: string
          description: Filter by locale (e.g., en, fr, es)
        - name: content
          in: query
          schema:
            type: string
          description: Search within translation values
        - name: tags
          in: query
          schema:
            type: array
            items:
              type: integer
          description: Filter by tag IDs
        - name: per_page
          in: query
          schema:
            type: integer
            default: 15
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranslationCollection"

  /api/tags:
    get:
      tags:
        - Tags
      summary: List all tags
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Tag"
    post:
      tags:
        - Tags
      summary: Create a new tag
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  example: mobile
      responses:
        "201":
          description: Created

  /api/tags/{id}:
    get:
      tags:
        - Tags
      summary: Get a tag by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
        "404":
          description: Not found
    put:
      tags:
        - Tags
      summary: Update a tag
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
      responses:
        "200":
          description: Success
        "404":
          description: Not found
    delete:
      tags:
        - Tags
      summary: Delete a tag
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content
        "404":
          description: Not found

  /api/export:
    get:
      tags:
        - Export
      summary: Export all translations as JSON
      description: |
        Returns all translations grouped by locale and key.
        Optimized for 100k+ records with response time under 500ms.
        Always returns fresh data (no caching).
      security:
        - bearerAuth: []
      parameters:
        - name: locale
          in: query
          schema:
            type: string
          description: Filter by specific locale
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email

    Translation:
      type: object
      properties:
        id:
          type: integer
        key:
          type: string
          example: common.welcome
        locale:
          type: string
          example: en
        value:
          type: string
          example: Welcome to our application
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    CreateTranslation:
      type: object
      required:
        - key
        - locale
        - value
      properties:
        key:
          type: string
          example: common.welcome
        locale:
          type: string
          example: en
        value:
          type: string
          example: Welcome to our application
        tags:
          type: array
          items:
            type: integer
          example: [1, 2]

    UpdateTranslation:
      type: object
      properties:
        key:
          type: string
        locale:
          type: string
        value:
          type: string
        tags:
          type: array
          items:
            type: integer

    TranslationCollection:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Translation"
        links:
          type: object
          properties:
            first:
              type: string
            last:
              type: string
            prev:
              type: string
              nullable: true
            next:
              type: string
              nullable: true
        meta:
          type: object
          properties:
            current_page:
              type: integer
            from:
              type: integer
            last_page:
              type: integer
            per_page:
              type: integer
            to:
              type: integer
            total:
              type: integer

    Tag:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          example: mobile
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    ExportResponse:
      type: object
      properties:
        data:
          type: object
          additionalProperties:
            type: object
            additionalProperties:
              type: string
          example:
            en:
              common.welcome: Welcome
              common.goodbye: Goodbye
            fr:
              common.welcome: Bienvenue
              common.goodbye: Au revoir